+++ scriptworker/config.py
@@ -43,8 +43,9 @@ DEFAULT_CONFIG = {
     "sign_chain_of_trust": False,  # TODO True
     "chain_of_trust_hash_algorithm": "sha256",
     "cot_schema_path": os.path.join(os.path.dirname(__file__), "data", "firefox_cot_schema.json"),
-    # Specify to override $HOME/.gnupg
+    # Specify a default gpg home other than ~/.gnupg
     "gpg_home": None,
+
     # A list of additional gpg cmdline options
     "gpg_options": None,
     # The path to the gpg executable.