+++ scriptworker/gpg.py
@@ -1,8 +1,10 @@
 #!/usr/bin/env python
-"""GPG functions
+"""GPG functions.  These currently assume gpg 2.0.x
 """
+import arrow
 import gnupg
 import logging
+import os
 
 from scriptworker.exceptions import ScriptWorkerTaskException
 
@@ -19,30 +21,54 @@ GPG_CONFIG_MAPPING = {
 }
 
 
-def GPG(context):
+def create_gpg_conf(homedir, my_fingerprint):
+    """ Set infosec guidelines; use my_fingerprint by default
+    """
+    gpg_conf = os.path.join(homedir, "gpg.conf")
+    if os.path.exists(gpg_conf):
+        os.rename(gpg_conf, "{}.{}".format(gpg_conf, arrow.utcnow().timestamp))
+    with open(gpg_conf, "w") as fh:
+        # https://wiki.mozilla.org/Security/Guidelines/Key_Management#GnuPG_settings
+        print("personal-digest-preferences SHA512 SHA384\n"
+              "cert-digest-algo SHA256\n"
+              "default-preference-list SHA512 SHA384 AES256 ZLIB BZIP2 ZIP Uncompressed\n"
+              "keyid-format 0xlong\n", file=fh)
+
+        # XXX If we want keyservers, we can specify, ideally configurable:
+        # print("keyserver gpg.mozilla.org\n"
+        #       "keyserver hkp://keys.gnupg.net\n"
+        #       "keyserver-options auto-key-retrieve\n", file=fh)
+
+        # default key
+        print("default-key {}".format(my_fingerprint), file=fh)
+
+
+def GPG(context, gpg_home=None):
     """Return a python-gnupg GPG instance
     """
     kwargs = {}
     for config_key, gnupg_key in GPG_CONFIG_MAPPING.items():
         if context.config[config_key] is not None:
             kwargs[gnupg_key] = context.config[config_key]
+    if gpg_home is not None:
+        kwargs['gnupghome'] = gpg_home
     gpg = gnupg.GPG(**kwargs)
     gpg.encoding = context.config['gpg_encoding'] or gpg.encoding
     return gpg
 
 
-def sign(context, data, **kwargs):
+def sign(context, data, gpg_home=None, **kwargs):
     """Sign `data` with the key `kwargs['keyid']`, or the default key if not specified
     """
-    gpg = GPG(context)
+    gpg = GPG(context, gpg_home=gpg_home)
     return str(gpg.sign(data, **kwargs))
 
 
-def verify_signature(context, signed_data, **kwargs):
+def verify_signature(context, signed_data, gpg_home=None, **kwargs):
     """Verify `signed_data` with the key `kwargs['keyid']`, or the default key if not specified
     """
     log.info("Verifying signature...")
-    gpg = GPG(context)
+    gpg = GPG(context, gpg_home=gpg_home)
     verified = gpg.verify(signed_data, **kwargs)
     if verified.trust_level is not None and verified.trust_level >= verified.TRUST_FULLY:
         log.info("Fully trusted signature from {}, {}".format(verified.username, verified.key_id))
@@ -51,10 +77,10 @@ def verify_signature(context, signed_data, **kwargs):
     return verified
 
 
-def get_body(context, signed_data, **kwargs):
+def get_body(context, signed_data, gpg_home=None, **kwargs):
     """Returned the unsigned data from `signed_data`.
     """
-    gpg = GPG(context)
+    gpg = GPG(context, gpg_home=gpg_home)
     verify_signature(context, signed_data)
     body = gpg.decrypt(signed_data, **kwargs)
     return str(body)