+++ CHANGELOG.md
@@ -3,8 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
+- ability to create new gpg homedirs
+
 ### Changed
 - `utils.makedirs` now throws `ScriptWorkerException` if the path exists and is not a directory or a softlink pointing to a directory.
+- gpg functions now take a `gpg_home` kwarg to specify a different homedir
 
 ## [0.5.0] - 2016-08-29
 ### Added